+++ src/conversion/utils/extractParameterNamesFromTask.ts
@@ -10,7 +10,7 @@ import { extractVariablesFromScript } from './extractVariablesFromScript';
  * @param task the task with used parameters
  * @returns the set of parameter names
  * @throws {ParseError} if the script is invalid
- * @public exported from `@promptbook/utils`
+ * @public exported from `@promptbook/core` <- Note: [👖] This utility is so tightly interconnected with the Promptbook that it is not exported as util but in core
  */
 export function extractParameterNamesFromTask(
     task: ReadonlyDeep<